Fix possible incorrect code tabe size
For images where the number of colors is smaller than the clear code,
there should be a gap between the default color codes and the clear
code. Otherwise new codes will be inserted at the wrong position and
further code lookups will fail with a NullpointerException. This has
been fixed.