Release 1.09
* Fixed: Dispose background is now handled properly and only clears the
area of the current graphic. Thanks to GitHub user wsxjump for reporting
this issue.
* Fixed: With wsxjump's initial fix sometimes image fragments were not
properly cleared. This should be fixed. Thanks to GitHub user tduva for
reporting.
* Fixed: A copy of every drawn frame is now kept in memory. These copies
won't be modified or used as input in subsequent frame processing. This
way no internal state is exposed and working with the frame images
returned by the decoder should be completely safe for clients. Thanks to
GitHub user MarkJeronimus for reporting this issue.
* Refactored various parts of the code.